Add ParameterSharingModule
