+++ tests/helpers/__init__.py
@@ -2,6 +2,7 @@ from tests.helpers.boring_model import (  # noqa: F401
     BoringDataModule,
     BoringModel,
     ManualOptimBoringModel,
+    ParameterSharingModule,
     RandomDataset,
 )
 from tests.helpers.datasets import TrialMNIST  # noqa: F401
