+++ tests/helpers/boring_model.py
@@ -14,6 +14,7 @@
 from typing import Optional
 
 import torch
+from torch import nn
 from torch.utils.data import DataLoader, Dataset, IterableDataset, Subset
 
 from pytorch_lightning import LightningDataModule, LightningModule
@@ -197,3 +198,18 @@ class ManualOptimBoringModel(BoringModel):
         self.manual_backward(loss)
         opt.step()
         return loss
+
+
+class ParameterSharingModule(BoringModel):
+    def __init__(self):
+        super().__init__()
+        self.layer_1 = nn.Linear(32, 10, bias=False)
+        self.layer_2 = nn.Linear(10, 32, bias=False)
+        self.layer_3 = nn.Linear(32, 10, bias=False)
+        self.layer_3.weight = self.layer_1.weight
+
+    def forward(self, x):
+        x = self.layer_1(x)
+        x = self.layer_2(x)
+        x = self.layer_3(x)
+        return x
