+++ tests/utilities/test_parameter_tying.py
@@ -16,22 +16,7 @@ import torch
 from torch import nn
 
 from pytorch_lightning.utilities import find_shared_parameters, set_shared_parameters
-from tests.helpers import BoringModel
-
-
-class ParameterSharingModule(BoringModel):
-    def __init__(self):
-        super().__init__()
-        self.layer_1 = nn.Linear(32, 10, bias=False)
-        self.layer_2 = nn.Linear(10, 32, bias=False)
-        self.layer_3 = nn.Linear(32, 10, bias=False)
-        self.layer_3.weight = self.layer_1.weight
-
-    def forward(self, x):
-        x = self.layer_1(x)
-        x = self.layer_2(x)
-        x = self.layer_3(x)
-        return x
+from tests.helpers import BoringModel, ParameterSharingModule
 
 
 @pytest.mark.parametrize(
